refactored small_World_functions.py into two files, one for bridge calculations and one for image generation. Renamed examples accordingly
